cleaned up plotting routines to make them more user friendly
