+++ pyMAP/tof.py
@@ -2,25 +2,29 @@ import numpy as np
 import pandas as pd
 import periodictable as perd
 
-# IBEX-lo Tof3 quadrant values, determined through calibration
+# IBEX-lo Tof3 quadrant values, determined through flight data analysis
 tof_3_peaks = [0.6900568049562702, 4.018218774092078, 7.277728085556459, 11.991842880356202]
+# IMAP-lo em quadrant values, determined through spectra analysis 
+tof3_peaks_EM = [ 1.92932145,  4.81973704,  7.84855093, 12.6265124 ]
 # electron flight time determined through simulation
 t_elec = {1:4.375828,2:4.097131}
 tof_dims = {'TOF0':50,'TOF1':22.5,'TOF2':27.7}
+m_elec = .00055 #AMU
 
-def tof_expected(ke_in=16000,
+
+def tof_expected(ke=16000,
                  species = 'H',
                  mass = None,
                 quadrant = 0,include_delay = False,q = 1,e_loss = 0):
 
-    ke = np.array(ke_in).reshape(-1)
+    ke = np.array(ke).reshape(-1)
     if type(species) is str:
         species = np.array(species.split(',')).reshape(-1)
     else:
         species = np.array(species).reshape(-1)
 
-    if mass is None:
-        mass = np.array([perd.elements.symbol(spec if spec != 'H2' else 'D').mass for spec in species]).reshape(-1)
+    if mass == None:
+        mass = np.array([(perd.elements.symbol(spec if spec != 'H2' else 'D').mass if spec != 'e' else m_elec) for spec in species]).reshape(-1)
     else: 
         mass = np.array(mass).reshape(-1)
         species = ['NaN']*len(mass)
@@ -54,12 +58,24 @@ def tof_expected(ke_in=16000,
     # dfs.columns = pd.MultiIndex.from_arrays([list(dfs.keys().values),units],names  =['','Units']) 
     return(dfs)
     
+def mass_line(ke):
+    return(pd.concat([tof_expected(ke,mass = m*perd.mass.neutron_mass) for m in range(32)]))
+
 def v_00(m,Vinc=7000,q = 1):
     # For a given mass and voltage drop, calculates the resulting velocity in cm/nS
     amu_c = 1.66*10**-27
     cm_c = 10**6
     qVinc = Vinc*1.6*10**-19*q
     return(np.sqrt(qVinc/m*2/amu_c)/cm_c)
+
+def tof_to_ke(tof,m,leg = 'TOF0',q = 1):
+    # For a given tof,mass and leg, calculates the incident energy in eV
+    amu_c = 1.66*10**-27
+    cm_c = 10**6
+    qv = 1.6*10**-19
+    return((tof_dims[leg]/tof*cm_c)**2*amu_c*m/(2*qv*q))
+    # return(np.sqrt(qVinc/m*2/amu_c)/cm_c)
+
 
 def delay_line_offset(tof3=tof_3_peaks):
 
@@ -93,36 +109,36 @@ def calc_eLoss(df):
     vs = tof_speeds(df)
     return(vs['TOF1']**2/vs['TOF2']**2)
 
-# def remove_delay_line(df_in):
-#     df = df_in.copy()      
-#     tof0 = df['TOF0']
-#     tof1 = df['TOF1']
-#     tof2 = df['TOF2']
-#     tof3 = df['TOF3']
+def remove_delay_line(df_in):
+    df = df_in.copy()      
+    tof0 = df['TOF0']
+    tof1 = df['TOF1']
+    tof2 = df['TOF2']
+    tof3 = df['TOF3']
     
-#     from scipy.interpolate import interp1d
-#     def delay_interp(tof3,d_effects):
-#         f_rn = interp1d(d_effects['tof3'],d_effects['Q'],kind = 'nearest',
-#                             bounds_error = False,fill_value="extrapolate")
-#         return(d_effects[['b0','b3']].iloc[f_rn(tof3).astype(int)]) 
+    from scipy.interpolate import interp1d
+    def delay_interp(tof3,d_effects):
+        f_rn = interp1d(d_effects['tof3'],d_effects['Q'],kind = 'nearest',
+                            bounds_error = False,fill_value="extrapolate")
+        return(d_effects[['b0','b3']].iloc[f_rn(tof3).astype(int)]) 
 
-#     # define the newly calculated tof values
-#     delay = delay_interp(tof3,delay_line_offset())
+    # define the newly calculated tof values
+    delay = delay_interp(tof3,delay_line_offset(tof3_peaks_EM))
 
-#     df['TOF0'] = tof0-delay['b0'].values +t_elec[1]
-#     df['TOF1'] = tof1-delay['b3'].values + t_elec[1]
-#     df['TOF2'] = tof2+t_elec[2] - t_elec[2]
+    df['TOF0'] = tof0-delay['b0'].values + t_elec[1]
+    df['TOF1'] = tof1-delay['b3'].values + t_elec[2]
+    df['TOF2'] = tof2+t_elec[1] - t_elec[2]
 
-#     return(df)
+    return(df)
 
 def get_checksum(df):
     return(calc_checksum(*[df['TOF%d'%i] for i in range(4)]))
 
-def remove_delay_line(df):
-    df_nd = df.copy()
-    df_nd['TOF0'] =df_nd['TOF0']+df_nd['TOF3']/2 
-    df_nd['TOF1'] =df_nd['TOF1']-df_nd['TOF3']/2 
-    return(df_nd)
+# def remove_delay_line(df):
+#     df_nd = df.copy()
+#     df_nd['TOF0'] =df_nd['TOF0']+df_nd['TOF3']/2 
+#     df_nd['TOF1'] =df_nd['TOF1']-df_nd['TOF3']/2 
+#     return(df_nd)
 
 def log_checksum(df,check_max = 1):
     return(abs(calc_checksum(*df[['TOF0','TOF1','TOF2','TOF3']].T.values))<check_max)
